refactor(search): split mail search into a package
Move EmailAddressIndex into suite/mail/search/indexes and keep the common
API (get/rebuild functions) in suite/mail/search/__init__.py. Public import
paths are unchanged, so callers and the whitelisted action need no edits.
