+++ suite/mail/search/__init__.py
@@ -1,79 +1,15 @@
-import re
-
 import frappe
 from frappe import _
 from frappe.utils import create_batch
 
-from suite.search.base_index import BaseIndex, FieldSpec
+from suite.mail.search.indexes import EmailAddressIndex
 from suite.utils import enqueue_job
-
-_TOKEN_PATTERN = re.compile(r"[a-z0-9]+")
 
 # Cached records processed per index write while rebuilding, bounding memory and commit size.
 _REBUILD_BATCH_SIZE = 500
 
 # Accounts rebuilt per long-queue job when rebuilding every account's index.
 _ACCOUNTS_PER_REBUILD_BATCH = 25
-
-
-class EmailAddressIndex(BaseIndex):
-	"""Shared, per-account index of email addresses for recipient suggestions.
-
-	Sources (cached messages, contact cards, ...) feed in plain {name, email} dicts, so the index
-	knows nothing about where an address came from. Each document is keyed by the lowercased
-	address, so re-indexing the same address from any source is an upsert and addresses stay unique
-	by construction. The index is cumulative: entries are only added or updated, never removed when
-	a source is evicted, so it doubles as an address book of everyone the user has corresponded with.
-	"""
-
-	ENTITY = "email_address"
-	FIELDS = (
-		# Lowercased address; the unique document key, so the same address upserts across sources.
-		FieldSpec("id", stored=True, tokenizer="raw"),
-		# Original-cased address and display name, returned verbatim in suggestions.
-		FieldSpec("email", stored=True, tokenizer="raw"),
-		FieldSpec("name", stored=True, tokenizer="raw"),
-		# "name email" blob, tokenized so a query can match either part.
-		FieldSpec("text"),
-	)
-	DEFAULT_SEARCH_FIELDS = ("text",)
-
-	def to_document(self, address: dict) -> dict:
-		email = (address.get("email") or "").strip()
-		name = (address.get("name") or "").strip() or None
-
-		return {
-			"id": email.lower(),
-			"email": email,
-			"name": name,
-			"text": " ".join(filter(None, (name, email))),
-		}
-
-	def index_addresses(self, addresses: list[dict]) -> int:
-		"""Upsert the given {name, email} dicts; skips entries without an email and dedupes the batch."""
-
-		unique = {}
-		for address in addresses:
-			if email := (address.get("email") or "").strip():
-				unique[email.lower()] = address
-
-		return self.index_documents(list(unique.values()))
-
-	def search_email_addresses(self, query: str, limit: int = 10) -> list[dict]:
-		"""Return up to `limit` {name, email} addresses matching `query`, most relevant first.
-
-		The query's tokens must appear as a consecutive, in-order phrase in the address's name or
-		email, with the last token matched as a prefix. So "saga" matches "sagar", and "sagar.s"
-		matches "sagar.s@…" / "Sagar Sharma" but not "sagar@…" or "sagar.v@…". Documents are unique
-		per address, so the hits need no further deduping.
-		"""
-
-		tokens = _TOKEN_PATTERN.findall(query.lower()) if query else []
-		if not tokens:
-			return []
-
-		hits, _total_count = self.search_phrase_prefix(tokens, limit=limit)
-		return [{"name": hit.get("name"), "email": hit.get("email")} for hit in hits]
 
 
 def get_email_address_index(account: str) -> EmailAddressIndex:
@@ -102,7 +38,7 @@ def rebuild_email_address_index(account: str, in_background: bool = True) -> Non
 		)
 		return
 
-	# Imported lazily: these modules import this one, so a top-level import would be circular.
+	# Imported lazily: these modules import this package, so a top-level import would be circular.
 	from suite.mail.doctype.contact_card.contact_card import _contact_addresses
 	from suite.mail.doctype.mail_message.mail_message import _message_addresses
 	from suite.mail.storage import get_data_store
