+++ suite/mail/search/indexes/__init__.py
@@ -1,0 +1,3 @@
+from suite.mail.search.indexes.email_address import EmailAddressIndex
+
+__all__ = ["EmailAddressIndex"]
